Fix bug where Grabber fails to download if config file points the non-existent download directory.
If Grabber is moved to other directory after configuration file is created, or for somewhat reason the directory where Grabber is set to download the images is renamed, Grabber is not able to handle such change and fails to download any images at all.

This behavior is fixed by making Grabber to check the directory for itself to download images into and to force-update the directory to $PWD if any error is detected.